# Auto-generated conversion methods calling one another