Part of #6800 

Conversion time decreased by:

    benchmark                                    old ns/op     new ns/op     delta
    BenchmarkPodConversion                       149268        30690         -79.44%
    BenchmarkNodeConversion                      60733         35517         -41.52%
    BenchmarkReplicationControllerConversion     135528        32236         -76.21%

Comparing with no conversion methods:

    benchmark                                    old ns/op     new ns/op     delta
    BenchmarkPodConversion                       388502        30690         -92.10%
    BenchmarkNodeConversion                      192973        35517         -81.59%
    BenchmarkReplicationControllerConversion     385409        32236         -91.64%

@smarterclayton @lavalamp - your predictions of 10X improvements were so close to what it gives :)

cc @smarterclayton @lavalamp @fgrzadkowski