@wojtek-t 2015-04-30
Note: Please do NOT merge #7560 and this PR at the same time - they need to be rebased before merging the other one.