@smarterclayton 2015-04-30
Woohoo!  What's the memory use look like?

----- Original Message -----
> Part of #6800
> 
> Conversion time decreased by:
> 
>     benchmark                                    old ns/op     new ns/op
>     delta
>     BenchmarkPodConversion                       149268        30690
>     -79.44%
>     BenchmarkNodeConversion                      60733         35517
>     -41.52%
>     BenchmarkReplicationControllerConversion     135528        32236
>     -76.21%
> 
> Comparing with no conversion methods:
> 
>     benchmark                                    old ns/op     new ns/op
>     delta
>     BenchmarkPodConversion                       388502        30690
>     -92.10%
>     BenchmarkNodeConversion                      192973        35517
>     -81.59%
>     BenchmarkReplicationControllerConversion     385409        32236
>     -91.64%
> 
> @smarterclayton @lavalamp - your predictions of 10X improvements were so
> close to what it gives :)
> 
> cc @smarterclayton @lavalamp @fgrzadkowski
> You can view, comment on, or merge this pull request online at:
> 
>   https://github.com/GoogleCloudPlatform/kubernetes/pull/7556
> 
> -- Commit Summary --
> 
>   * Chain conversion functions while generation
>   * Update conversion functions
> 
> -- File Changes --
> 
>     M pkg/api/v1beta3/conversion.go (1603)
>     M pkg/conversion/converter.go (27)
>     M pkg/conversion/generator.go (79)
> 
> -- Patch Links --
> 
> https://github.com/GoogleCloudPlatform/kubernetes/pull/7556.patch
> https://github.com/GoogleCloudPlatform/kubernetes/pull/7556.diff
> 
> ---
> Reply to this email directly or view it on GitHub:
> https://github.com/GoogleCloudPlatform/kubernetes/pull/7556
>