@wojtek-t 2015-04-30
>
Woohoo!  What's the memory use look like?

When I checked it in the morning the gain was significant but I can't remember it now. I will rerun the experiments tomorrow (I need to run now).

BTW PR rebased