@smarterclayton 2015-04-30
Can we move the generated functions to their own file conversion_generated.go?