@lavalamp 2015-05-01
Awesome-- +1 for moving to a new file. I'll give a more detailed look later.