@wojtek-t 2015-05-01
>
Can we move the generated functions to their own file conversion_generated.go?

I agree that's a good idea. However, since today is a public holiday in Poland I might not find time to do it before Monday.