@wojtek-t 2015-05-04
Regarding memory, here is the comparison between having no generated conversions and this PR id ~3.5X better - the graphs below:

Before:
![profile_old](https://cloud.githubusercontent.com/assets/10743879/7450102/4f899a5e-f241-11e4-9287-6257bcc6e405.png)

After:
![profile_new](https://cloud.githubusercontent.com/assets/10743879/7450106/55a1dbf4-f241-11e4-8193-7385f5ed10b0.png)

Also, I've updated PR so that the generated code is in a separate file.

PTAL