@wojtek-t 2015-05-04
BTW, once we agree that this is the final format, I will generate conversions to v1 api version in a separate PR.