@wojtek-t 2015-05-04
**Update on memory**
I think the the gain is bigger than 3.5X that i mentioned. Basically without generated conversion functions, there are 900 repetitions of methods, and with them generated 14.000 repetitions. So I think that we should multiply this 3.5X by 15 :) [or sth like that].