@lavalamp 2015-05-04
So my major concern here is how does this interact with human defined conversion functions? When we make an API change, can I still write a conversion function like I could before? Or another way of putting it, can we turn this on for v1beta1/2?