@wojtek-t 2015-05-04
>
So my major concern here is how does this interact with human defined conversion functions? When we make an API change, can I still write a conversion function like I could before? Or another way of putting it, can we turn this on for v1beta1/2?

You definitely should be able to define the conversion function yourself (currently the conversion function wouldn't be generated if it's not obvious how to translate it - i.e. if there are field with different names, or of types that we couldn't automatically detect how to convert). In such cases, writing your own conversion function is obvious (it wouldn't conflict with anything)
The only question is what happens if translation could be detected automatically but we want the conversion to be done in a different way. But I think that we should block this PR on it.