@lavalamp 2015-05-04
I see. I was hoping we could read the manually written conversion functions, and use them to generate the expanded conversion functions.

> The only question is what happens if translation could be detected automatically but we want the conversion to be done in a different way. But I think that we should block this PR on it.

I think this effectively will prevent us from mutating our api. We'll need a solution for this soon. I think we can still merge this if you have a plan to fix this soon.