@smarterclayton 2015-05-05
----- Original Message -----
> >
> I see. I was hoping we could read the manually written conversion functions,
> and use them to generate the expanded conversion functions.
> 
> My plan here is to change the manually written conversion functions to be
> named (with some explicitly defined naming convention) and the use them
> while generating conversion functions. I just didn't get to it yet.
> 
> >
> I think this effectively will prevent us from mutating our api. We'll need a
> solution for this soon. I think we can still merge this if you have a plan
> to fix this soon.
> 
> Yes, I thought about it and I think it should be possible to do it pretty
> soon. Basically my plan is to:
> 1. internally, store the generated and manually-written conversion functions
> separately
> 2. in default (reflection-based) implementation, check whether any of those
> exists
> 3. while generating conversion methods, if there is a manually-written
> conversion function, do NOT generate one for it

Would it be an expectation on the manual-written conversion author to invoke the generated function by name?  I.e.:

func myconversion() {
  out.Bar = in.Bar
  convert_api_SomeObject_to_v1beta3_SomeObject(out, in)
}

?  If so, that seems reasonable.