@wojtek-t 2015-05-05
>
Would it be an expectation on the manual-written conversion author to invoke the generated function by name?  I.e.:

func myconversion() {
  out.Bar = in.Bar
  convert_api_SomeObject_to_v1beta3_SomeObject(out, in)
}

?  If so, that seems reasonable.

That's exactly what I thought about.