@lavalamp 2015-05-05
LGTM

Another future improvement would be to inline the defaulting code. It's making a lot of reflect.TypeOf calls right now.