@wojtek-t 2015-05-06
>
Another future improvement would be to inline the defaulting code. It's making a lot of reflect.TypeOf calls right now.

Sure - but that's just an optimization.